Put book info views into separate file
